# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Quickstart in Couchbase with PHP and Laravel",
...
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/v1/routes/list": {
            "get": {
                "tags": [
                    "Routes"
                ],
                "summary": "Get list of routes",
                "description": "Returns list of routes",
                "operationId": "getRoutesList",
                "parameters": [
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 0
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 10
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Route"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/routes/{id}": {
            "get": {
                "tags": [
                    "Routes"
                ],
                "summary": "Get route information",
                "description": "Returns route data",
                "operationId": "getRouteById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Route"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Routes"
                ],
                "summary": "Update an existing route or create a new one if it does not exist",
                "description": "Update an existing route or create a new one if it does not exist",
                "operationId": "updateRoute",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Route object that needs to be updated or created",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Route"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "201": {
                        "description": "Resource created successfully"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Routes"
                ],
                "summary": "Create a new route",
                "description": "Create a new route",
                "operationId": "createRoute",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Route object that needs to be added",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Route"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Routes"
                ],
                "summary": "Delete a route",
                "description": "Delete a route",
                "operationId": "deleteRoute",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Route": {
                "title": "Route",
                "required": [
                    "airline",
                    "airlineid",
                    "sourceairport",
                    "destinationairport",
                    "stops",
                    "equipment",
                    "schedule",
                    "distance"
                ],
                "properties": {
                    "airline": {
                        "description": "Airline code",
                        "type": "string"
                    },
                    "airlineid": {
                        "description": "Airline ID",
                        "type": "string"
                    },
                    "sourceairport": {
                        "description": "Source airport code",
                        "type": "string"
                    },
                    "destinationairport": {
                        "description": "Destination airport code",
                        "type": "string"
                    },
                    "stops": {
                        "description": "Number of stops",
                        "type": "integer"
                    },
                    "equipment": {
                        "description": "Equipment used",
                        "type": "string"
                    },
                    "schedule": {
                        "description": "Flight schedule",
                        "type": "array",
                        "items": {
                            "properties": {
                                "day": {
                                    "description": "Day of the flight",
                                    "type": "integer"
                                },
                                "utc": {
                                    "description": "UTC time of the flight",
                                    "type": "string"
                                },
                                "flight": {
                                    "description": "Flight number",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "distance": {
                        "description": "Distance of the route",
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            }
        }
    }
}